Footprint: add method to shrink, removing edge masked pixels
Currently, two sources on either side of a chip or exposure gap on a coadd
are joined by high (interpolated) pixels masked NO_DATA.  Shrinking allows
us to remove those meaningless pixels from the Footprint.  We prefer this
to ignoring the mask in the initial thresholding that created the
Footprints because this shrinking (from the outside in) will not unselect
NO_DATA pixels in the core of a star.