add findFootprintAtPoint
This is used in culling peaks, and needs to be in C++ for speed
reasons.  The original implementation was written by Robert Lupton
for Pan-STARRS in C, and slightly-ported to LSST C++ a long time
ago; I've cleaned up and finished this work.

Also includes checkFootprintAtPoint, which does not return the
full Footprint, but only whether the Footprint includes a stopping
point (as is required for culling peaks).